add sonar sensor for down stairs, pb_teleop_sub receives the topics from cpp, moving flipper with thread, Now, it can perform "going up stairs" and "going down stairs".  Changes to be committed: 	modified:   jaden/jaden-pairbot/readme.md 	modified:   jaden/jaden-pairbot/src/pairbot/pb_description/urdf/pairbot.gazebo 	modified:   jaden/jaden-pairbot/src/pairbot/pb_description/urdf/pairbot.xacro 	modified:   jaden/jaden-pairbot/src/pairbot/pb_teleop/scripts/pb_teleop_sub.py 	modified:   jaden/jaden-pairbot/src/stair_climbing/launch/down_stair.launch 	modified:   jaden/jaden-pairbot/src/stair_climbing/launch/stair_mark.launch 	modified:   jaden/jaden-pairbot/src/stair_climbing/src/DownStairMode.cpp 	modified:   jaden/jaden-pairbot/src/stair_climbing/src/DownStairMode.h 	modified:   jaden/jaden-pairbot/src/stair_climbing/src/Stair_Det.cpp 	modified:   jaden/jaden-pairbot/src/stair_climbing/src/StairsDet.cpp 	modified:   jaden/jaden-pairbot/src/stair_climbing/src/StairsDet.h 	modified:   jaden/jaden-pairbot/src/stair_climbing/src/UpStairMode.cpp

+++ jaden/jaden-pairbot/src/stair_climbing/src/StairsDet.h
@@ -33,7 +33,7 @@ class StairDetection
 			d_sub = it.subscribe("/Zed_sim/depth/depth_registered", 10, &StairDetection::imageCallback,this);
 			rgb_sub=it.subscribe("/zed/rgb/image_rect_color", 1, &StairDetection::rgbCallback,this);
 			pub_d21 = n.advertise<std_msgs::Int32>("d21", 1);
-			pub_cnt21 = n.advertise<std_msgs::Float64>("cnt21", 1);
+			pub_cnt21 = n.advertise<std_msgs::Int32>("cnt21", 1);
 			pub_d4 = n.advertise<std_msgs::Int32>("d4", 1);
 			pub_mid = n.advertise<std_msgs::Int32MultiArray>("midxy", 1);
 
@@ -49,7 +49,7 @@ class StairDetection
 
 		void test();
 		void Depth_Arr();
-		float depthMD(int x,int y);
+		int depthMD(int x,int y);
 		void depth_find(int dc,int *mid_x2,int *mid_y2,int *midx,int *midy,int *di,int *dj,int *dx,int *dsw);
 		void depth_find_4(int dc,int *midx,int *midy,int *di,int *dj,int *dx,int *dsw);
 		void depth_find_5(int dc,int *midx,int *midy,int *di,int *dj,int *dx,int *dsw);
